requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.26
langsmith>=0.0.87
qdrant-client>=1.7.0
streamlit>=1.29.0
pypdf>=3.17.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
openai>=1.6.0
tiktoken>=0.5.0

